refactor(skia-treemap): hoist into _Shared source-gen pipeline
Mirror the PieChart layout so the SK headless chart and per-platform
view controls all consume the same partial class:

- _Shared/SourceGenTreemapChart.shared.cs: shared partial with
  CreateCoreChart + InitializeObservedProperties (#if SKIA_IMAGE_LVC
  switches the class name to SourceGenSKTreemapChart for the headless
  build, SourceGenTreemapChart for the per-platform builds).
- _Shared/SourceGenTreemapChart.sgp.cs: placeholder for chart-level
  UIProperty markers (Treemap has none yet — every configurable property
  lives on the series).
- _Shared.WinUI/SourceGenTreemapChart.winui.cs: WinUI/Uno-specific
  partial mirroring SourceGenPieChart.winui.cs.
- SKCharts/LiveChartsGeneratedCode/SourceGenSKTreemapChart.cs: SK
  headless base (constructor accepting IChartView?) mirroring
  SourceGenSKPieChart.cs.
- SKTreemapChart.cs: now a thin two-constructor wrapper, prior
  hand-rolled IChartView impl moved into the shared partial.

Wired into _Shared.projitems (consumed by the SK project) +
SharedLinks.Build.props + WinUISharedLinks.Build.props (consumed by the
per-platform projects).
