+++ src/skiasharp/_Shared.WinUI/_Shared.WinUI.projitems
@@ -21,6 +21,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)SourceGenMapChart.winui.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SourceGenPieChart.winui.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SourceGenPolarChart.winui.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SourceGenTreemapChart.winui.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)TreemapChart.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ThemeListener.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)XamlSeries.cs" />
   </ItemGroup>
