+++ src/skiasharp/_Shared/_Shared.projitems
@@ -16,5 +16,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)SourceGenPieChart.sgp.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SourceGenPolarChart.shared.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SourceGenPolarChart.sgp.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SourceGenTreemapChart.shared.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SourceGenTreemapChart.sgp.cs" />
   </ItemGroup>
 </Project>
